Reorganize dataset module, adding base module, and add model-based CBMA methods.
- Move Analyzable, Mappable, ConnMatrix, Image, ActivationSet, and
Surface from dataset/base to base/inputs.
- Move Contrast from dataset/dataset to dataset/base.
- Add module docstrings for decoding module.
- Add skeleton for model-based CBMA methods (BHICP, HPGRF, SBLFR, SBR).
- Add duecredit citations for MKDA and model-based CBMA methods.
