+++ nimare/__init__.py
@@ -1,7 +1,8 @@
 # -*- coding: utf-8 -*-
 # emacs: -*- mode: python; py-indent-offset: 4; indent-tabs-mode: nil -*-
 # vi: set ft=python sts=4 ts=4 sw=4 et:
-""" NiMARE
+"""
+NiMARE: Neuroimaging Meta-Analysis Research Environment
 """
 from .dataset import Dataset
 from .meta import ALE
