+++ nimare/base/__init__.py
@@ -2,6 +2,7 @@
 # emacs: -*- mode: python; py-indent-offset: 4; indent-tabs-mode: nil -*-
 # vi: set ft=python sts=4 ts=4 sw=4 et:
 """
-Methods for decoding subsets of voxels (e.g., ROIs) or experiments (e.g., from
-meta-analytic clustering on a database) into text.
+Top-level namespace for nimare base.
 """
+
+from .inputs import Analyzable, Mappable, ConnMatrix, Image, ActivationSet, Surface
