+++ nimare/dataset/__init__.py
@@ -4,5 +4,5 @@
 """
 Top-level namespace for datasets.
 """
-
-from .dataset import Dataset, Contrast
+from .base import Contrast
+from .dataset import Dataset
