+++ nimare/decode/__init__.py
@@ -2,7 +2,7 @@
 # emacs: -*- mode: python; py-indent-offset: 4; indent-tabs-mode: nil -*-
 # vi: set ft=python sts=4 ts=4 sw=4 et:
 """
-Top-level namespace for meta-analyses.
+Top-level namespace for functional decoding.
 """
 
 from .continuous import GCLDAContinuousDecoder
