+++ nimare/decode/continuous.py
@@ -1,0 +1,6 @@
+# -*- coding: utf-8 -*-
+# emacs: -*- mode: python; py-indent-offset: 4; indent-tabs-mode: nil -*-
+# vi: set ft=python sts=4 ts=4 sw=4 et:
+"""
+Methods for decoding unthresholded brain maps into text.
+"""
