+++ nimare/meta/cbma/__init__.py
@@ -7,3 +7,4 @@ Top-level namespace for meta-analyses.
 
 from .ale import ALE, SCALE
 from .mkda import MKDA
+from .model import BHICP, HPGRF, SBLFR, SBR
